empresas-automatizadas-ia/exercicio-13-dspy-optimizer/requirements.txt
python-dotenv>=1.0.0,<2
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
pydantic>=2.10.0,<3
dspy-ai>=2.5.0,<3
